4 Star Code Gacha :O

+++ Library/Larva.cpp
@@ -3,7 +3,7 @@
 #include "Bees.h"
 
 std::default_random_engine generator;
-std::normal_distribution<double> distribution(11.0, 5.0);
+std::normal_distribution<float> distribution(11.0, 5.0);
 const float Larva::EMERGE_TIME_QUEEN{ 24 * distribution(generator) / TIME_SCALING };
 const float Larva::EMERGE_TIME_DRONE{ 24 * distribution(generator) / TIME_SCALING };
 const float Larva::EMERGE_TIME_WORKER{ 24 * distribution(generator) / TIME_SCALING };
